[feature] add xmldb:find-last-modified-until
