+++ src/org/exist/xquery/functions/xmldb/XMLDBModule.java
@@ -88,7 +88,8 @@ public class XMLDBModule extends AbstractInternalModule {
             new FunctionDef(XMLDBReindex.FNS_REINDEX_DOCUMENT, XMLDBReindex.class),
             new FunctionDef(XMLDBDefragment.signatures[0], XMLDBDefragment.class),
             new FunctionDef(XMLDBDefragment.signatures[1], XMLDBDefragment.class),
-            new FunctionDef(FindLastModifiedSince.signature, FindLastModifiedSince.class),
+            new FunctionDef(FindLastModified.signatures[0], FindLastModified.class),
+            new FunctionDef(FindLastModified.signatures[1], FindLastModified.class),
             new FunctionDef(XMLDBMatchCollection.signature, XMLDBMatchCollection.class),
             new FunctionDef(XMLDBTouch.FNS_TOUCH_DOCUMENT_NOW, XMLDBTouch.class),
             new FunctionDef(XMLDBTouch.FNS_TOUCH_DOCUMENT, XMLDBTouch.class),
